+++ src/main/java/com/epam/aidial/deployment/manager/service/manifest/ProbeConverter.java
@@ -13,7 +13,7 @@ import org.jetbrains.annotations.Nullable;
 import org.springframework.stereotype.Component;
 
 /**
- * Converts domain probe properties to Kubernetes probe and to KServe/NIM startup probe types.
+ * Converts domain probe properties to Kubernetes probe.
  */
 @Slf4j
 @Component
@@ -110,136 +110,4 @@ public class ProbeConverter {
             @Nullable IntOrString httpPort,
             @Nullable String httpPath
     ) { }
-
-    /**
-     * Converts domain probe properties to KServe predictor model StartupProbe.
-     *
-     * @param probeProperties domain probe properties, or null
-     * @return KServe StartupProbe or null if probe is null
-     */
-    @Nullable
-    public io.kserve.serving.v1beta1.inferenceservicespec.predictor.model.StartupProbe toKserveStartupProbe(
-            @Nullable ProbeProperties probeProperties) {
-        var probe = toProbe(probeProperties);
-        return toKserveStartupProbe(probe);
-    }
-
-    /**
-     * Converts fabric8 Probe to KServe predictor model StartupProbe.
-     *
-     * @param probe fabric8 Probe, or null
-     * @return KServe StartupProbe or null if probe is null
-     */
-    @Nullable
-    private io.kserve.serving.v1beta1.inferenceservicespec.predictor.model.StartupProbe toKserveStartupProbe(
-            @Nullable Probe probe) {
-        if (probe == null) {
-            log.debug("Probe is null, skipping KServe startup probe conversion");
-            return null;
-        }
-        var kserve = new io.kserve.serving.v1beta1.inferenceservicespec.predictor.model.StartupProbe();
-        copyTimingToKserve(probe, kserve);
-        copyHandlerToKserve(probe, kserve);
-        log.debug("Converted fabric8 Probe to KServe StartupProbe");
-        return kserve;
-    }
-
-    private static void copyTimingToKserve(Probe source,
-            io.kserve.serving.v1beta1.inferenceservicespec.predictor.model.StartupProbe target) {
-        if (source.getInitialDelaySeconds() != null) {
-            target.setInitialDelaySeconds(source.getInitialDelaySeconds());
-        }
-        if (source.getPeriodSeconds() != null) {
-            target.setPeriodSeconds(source.getPeriodSeconds());
-        }
-        if (source.getTimeoutSeconds() != null) {
-            target.setTimeoutSeconds(source.getTimeoutSeconds());
-        }
-        if (source.getFailureThreshold() != null) {
-            target.setFailureThreshold(source.getFailureThreshold());
-        }
-        if (source.getSuccessThreshold() != null) {
-            target.setSuccessThreshold(source.getSuccessThreshold());
-        }
-        if (source.getTerminationGracePeriodSeconds() != null) {
-            target.setTerminationGracePeriodSeconds(source.getTerminationGracePeriodSeconds());
-        }
-    }
-
-    private static void copyHandlerToKserve(Probe source,
-            io.kserve.serving.v1beta1.inferenceservicespec.predictor.model.StartupProbe target) {
-        if (source.getHttpGet() != null) {
-            var httpGet = new io.kserve.serving.v1beta1.inferenceservicespec.predictor.model.startupprobe.HttpGet();
-            httpGet.setPath(source.getHttpGet().getPath());
-            httpGet.setPort(source.getHttpGet().getPort());
-            target.setHttpGet(httpGet);
-        }
-    }
-
-    /**
-     * Converts domain probe properties to NIM StartupProbe (with enabled set to true and probe payload).
-     *
-     * @param probeProperties domain probe properties, or null
-     * @return NIM StartupProbe or null if probe is null
-     */
-    @Nullable
-    public com.nvidia.apps.v1alpha1.nimservicespec.StartupProbe toNimStartupProbe(
-            @Nullable ProbeProperties probeProperties) {
-        var probe = toProbe(probeProperties);
-        return toNimStartupProbe(probe);
-    }
-
-    /**
-     * Converts fabric8 Probe to NIM StartupProbe (with enabled set to true and probe payload).
-     *
-     * @param probe fabric8 Probe, or null
-     * @return NIM StartupProbe or null if probe is null
-     */
-    @Nullable
-    private com.nvidia.apps.v1alpha1.nimservicespec.StartupProbe toNimStartupProbe(@Nullable Probe probe) {
-        if (probe == null) {
-            log.debug("Probe is null, skipping NIM startup probe conversion");
-            return null;
-        }
-        var nimStartupProbe = new com.nvidia.apps.v1alpha1.nimservicespec.StartupProbe();
-        nimStartupProbe.setEnabled(true);
-        var nimProbe = new com.nvidia.apps.v1alpha1.nimservicespec.startupprobe.Probe();
-        copyTimingToNimProbe(probe, nimProbe);
-        copyHandlerToNimProbe(probe, nimProbe);
-        nimStartupProbe.setProbe(nimProbe);
-        log.debug("Converted fabric8 Probe to NIM StartupProbe");
-        return nimStartupProbe;
-    }
-
-    private static void copyTimingToNimProbe(Probe source,
-            com.nvidia.apps.v1alpha1.nimservicespec.startupprobe.Probe target) {
-        if (source.getInitialDelaySeconds() != null) {
-            target.setInitialDelaySeconds(source.getInitialDelaySeconds());
-        }
-        if (source.getPeriodSeconds() != null) {
-            target.setPeriodSeconds(source.getPeriodSeconds());
-        }
-        if (source.getTimeoutSeconds() != null) {
-            target.setTimeoutSeconds(source.getTimeoutSeconds());
-        }
-        if (source.getFailureThreshold() != null) {
-            target.setFailureThreshold(source.getFailureThreshold());
-        }
-        if (source.getSuccessThreshold() != null) {
-            target.setSuccessThreshold(source.getSuccessThreshold());
-        }
-        if (source.getTerminationGracePeriodSeconds() != null) {
-            target.setTerminationGracePeriodSeconds(source.getTerminationGracePeriodSeconds());
-        }
-    }
-
-    private static void copyHandlerToNimProbe(Probe source,
-            com.nvidia.apps.v1alpha1.nimservicespec.startupprobe.Probe target) {
-        if (source.getHttpGet() != null) {
-            var httpGet = new com.nvidia.apps.v1alpha1.nimservicespec.startupprobe.probe.HttpGet();
-            httpGet.setPath(source.getHttpGet().getPath());
-            httpGet.setPort(source.getHttpGet().getPort());
-            target.setHttpGet(httpGet);
-        }
-    }
 }
